donee

+++ Inception/srcs/requirements/wordpress/tools/script.sh
@@ -1,26 +1,28 @@
 #!/bin/bash
-mkdir /var/www/
-mkdir /var/www/html
-cd /var/www/html
-rm -rf *
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
-wp core download --allow-root
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-mv /wp-config.php /var/www/html/wp-config.php
-curl -s https://api.wordpress.org/secret-key/1.1/salt/ >> /var/www/html/wp-config.php
-echo "if ( ! defined( 'ABSPATH' ) ) {" >> /var/www/html/wp-config.php
-echo "    define( 'ABSPATH', '/var/www/html/' );" >> /var/www/html/wp-config.php
-echo "}" >> /var/www/html/wp-config.php
-echo "" >> /var/www/html/wp-config.php # Add an empty line for readability
-echo "require_once ABSPATH . 'wp-settings.php';" >> /var/www/html/wp-config.php
-sed -i -r "s/db1/$db_name/1"   wp-config.php
-sed -i -r "s/user/$db_user/1"  wp-config.php
-sed -i -r "s/pwd/$db_pwd/1"    wp-config.php
-wp core install --url=$DOMAIN_NAME/ --title=$WP_TITLE --admin_user=$WP_ADMIN_USR --admin_password=$WP_ADMIN_PWD --admin_email=$WP_ADMIN_EMAIL --skip-email --allow-root
-wp user create $WP_USR $WP_EMAIL --role=author --user_pass=$WP_PWD --allow-root
-wp plugin update --all --allow-root
+if [ ! -f /var/www/html/wp-config.php ]; then
+    mkdir /var/www/
+    mkdir /var/www/html
+    cd /var/www/html
+    rm -rf *
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+    wp core download --allow-root
+    mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+    mv /wp-config.php /var/www/html/wp-config.php
+    curl -s https://api.wordpress.org/secret-key/1.1/salt/ >> /var/www/html/wp-config.php
+    echo "if ( ! defined( 'ABSPATH' ) ) {" >> /var/www/html/wp-config.php
+    echo "    define( 'ABSPATH', '/var/www/html/' );" >> /var/www/html/wp-config.php
+    echo "}" >> /var/www/html/wp-config.php
+    echo "" >> /var/www/html/wp-config.php # Add an empty line for readability
+    echo "require_once ABSPATH . 'wp-settings.php';" >> /var/www/html/wp-config.php
+    sed -i -r "s/db1/$db_name/1"   wp-config.php
+    sed -i -r "s/user/$db_user/1"  wp-config.php
+    sed -i -r "s/pwd/$db_pwd/1"    wp-config.php
+    wp core install --url=$DOMAIN_NAME/ --title=$WP_TITLE --admin_user=$WP_ADMIN_USR --admin_password=$WP_ADMIN_PWD --admin_email=$WP_ADMIN_EMAIL --skip-email --allow-root
+    wp user create $WP_USR $WP_EMAIL --role=author --user_pass=$WP_PWD --allow-root
+    wp plugin update --all --allow-root
+    mkdir /run/php
+fi
 sed -i 's/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/g' /etc/php/7.3/fpm/pool.d/www.conf
-mkdir /run/php
 /usr/sbin/php-fpm7.3 -F
